LPS-178068 Rename service
