+++ workspaces/liferay-partner-workspace/client-extensions/liferay-partner-oauth-application-headless-server/src/main/java/com/liferay/partner/service/SalesforceService.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.partner.services;
+package com.liferay.partner.service;
 
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
